Refactor. yfhy

+++ src/main/java/com/thoughtworks/rectangle/Rectangle.java
@@ -1,19 +1,16 @@
 package com.thoughtworks.rectangle;
 
 public class Rectangle {
-    private final int length;
-    private final int breadth;
+    private final double length;
+    private final double breadth;
 
-
-    public Rectangle(int length, int breadth) {
+    public Rectangle(double length, double breadth) {
         this.length = length;
-
         this.breadth = breadth;
-
     }
 
-
-    public int area() {
+    public double area() {
         return this.length * this.breadth;
     }
+
 }
